address lint warnings
